add with robot in methods, modify optimization, change sqrt(dot()) to linalg.norm, change cone tolerance verification method, create ikfast_5d method
- add with robot: with statement save previous state of the robot and
when the function finishes, the robot returns to the previous state.
- modify optimization : not recursive anymore. The optimization goes
always forward an try backtracking if it fails on a point. If
backtracking fails, it will not check multiple branches, the
optimization returns the last achieved robot joints.
- change cone tolerance verification method:
the method inverse_kinematics_maximum_tolerance_angle made a cone with
the given tolerance. The chosen discretization could find a solution in
a first moment, and could not find in a secondary moment (as the
perpendicular vector is randomly selected). The solution is to find at
least two tolerance solutions in the cone and increase the
discretization from 8 to 72.